modified:   src/components/Header/Header.jsx 	modified:   src/components/InputBalance/InputBalance.jsx 	modified:   src/components/Layout/Layout.jsx 	modified:   src/components/Login/Login.jsx     modified:   src/components/ModalLoader/ModalLoader.jsx 	modified:   src/components/ModalLoader/ModalLoader.module.css 	modified:   src/components/Register/Register.jsx 	renamed:    src/redux/authSlice.js -> src/redux/authorization/authSlice.js 	renamed:    src/redux/operations.js -> src/redux/authorization/operations.js 	deleted:    src/redux/selectors.js 	modified:   src/redux/store.js 	renamed:    src/redux/constants.js -> src/redux/user/userActions.js 	renamed:    src/redux/userSlice.js -> src/redux/user/userSlice.js
