+++ src/components/ModalLoader/ModalLoader.module.css
@@ -1,6 +1,14 @@
 .modal {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background: rgba(0, 0, 0, 0.5);
   display: flex;
-  place-items: center;
-  width: 100%;
-  height: 100%;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+ 
 }
+
