+++ src/redux/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
-import authReducer from './authSlice';
-import userReducer from './userSlice';
+import authReducer from './authorization/authSlice';
+import userReducer from './user/userSlice';
 import categoriesReducer from './categories/categoriesSlice';
 import periodReducer from './period/periodSlice';
 import transactionReducer from './transaction/transactionSlice';
